test: polish node custom predicate tests
